Add TreeOptions.DoubleOddNode
Normally, when an odd node exists, it is simply forwarded to the next
layer. Now, if DoubleOddNode is true, then the odd node is doubled, and
becomes a left/right pair with itself. Thus it is concatenated, and
hashed to form a node in the next layer, just like any other pair.